requirements.txt
lxml
selenium
Flask
requests
flask-cors
python-dotenv==1.0.1
